backport preprocessing optimizations
use float2half_noLUT instead of float2half (on x86; bit identical),
use SSE optimized float2half-loop (on x86; bit identical),
use copy_bgra_rgba in DX system texture path,
use SSE optimized min_max-loop for RGB16F detection/handling

+++ src/parts/ball.cpp
@@ -233,12 +233,12 @@ void BallEx::Render(const unsigned int renderMask)
       if (!invalid)
       {
          // compute size of the rendered ball on viewport, then apply reversed viewport rotation, then compute stretch correction
-         const float w = m_rd->GetCurrentRenderTarget()->GetWidth();
-         const float h = m_rd->GetCurrentRenderTarget()->GetHeight();
+         const int w = m_rd->GetCurrentRenderTarget()->GetWidth();
+         const int h = m_rd->GetCurrentRenderTarget()->GetHeight();
          const float viewportRot = -ANGTORAD(g_pplayer->m_ptable->mViewSetups[g_pplayer->m_ptable->m_BG_current_set].GetRotation(w, h));
          const float c = cosf(viewportRot), s = sinf(viewportRot);
-         const float rx = (xMax - xMin) * w;
-         const float ry = (yMax - yMin) * h;
+         const float rx = (xMax - xMin) * (float)w;
+         const float ry = (yMax - yMin) * (float)h;
          const float sx = fabs(c * rx - s * ry);
          const float sy = fabs(s * rx + c * ry);
          // only shrink ball to avoid artefact of the ball being rendered over resting parts
